Move shop mapper to a separate file
