+++ app/src/main/java/com/simply/birthdayapp/data/shops/ShopsRepository.kt
@@ -1,6 +1,5 @@
-package com.simply.birthdayapp.data.repositories
+package com.simply.birthdayapp.data.shops
 
-import android.util.Base64
 import com.apollographql.apollo3.ApolloClient
 import com.simply.birthdayapp.ShopsQuery
 import com.simply.birthdayapp.presentation.ui.models.Shop
@@ -24,19 +23,4 @@ class ShopsRepositoryImpl(
             ?.map(ShopsQuery.Shop::toShop)
             ?: emptyList()
     }
-}
-
-private fun ShopsQuery.Shop.toShop(): Shop {
-    var imageData = byteArrayOf()
-    try {
-        imageData = Base64.decode(image.substringAfter("base64,"), Base64.DEFAULT)
-    } catch (cause: IllegalArgumentException) {
-        cause.printStackTrace()
-    }
-    return Shop(
-        id = id,
-        name = name,
-        image = imageData,
-        isFavorite = isFavorite,
-    )
 }
