+++ app/src/main/java/com/simply/birthdayapp/di/DataModule.kt
@@ -1,8 +1,8 @@
 package com.simply.birthdayapp.di
 
 import com.simply.birthdayapp.data.createApolloClient
-import com.simply.birthdayapp.data.repositories.ShopsRepository
-import com.simply.birthdayapp.data.repositories.ShopsRepositoryImpl
+import com.simply.birthdayapp.data.shops.ShopsRepository
+import com.simply.birthdayapp.data.shops.ShopsRepositoryImpl
 import org.koin.core.module.Module
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
